backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.6
httpx==0.26.0
aiosqlite==0.19.0
jinja2==3.1.3
email-validator==2.1.0
fpdf2==2.8.3
